+++ src/event/active-element.js
@@ -1,7 +1,10 @@
 
 /*
-  USAGE:
+  Debugging tool that observe changes to activeElement regardless of focus/blur events.
+  This utility does *not* work with Shadow DOM.
 
+  USAGE:
+    engage();
     document.body.addEventListener('active-element', function(event) {
       // event.detail.focus: element that received focus
       // event.detail.blur: element that lost focus
@@ -13,17 +16,18 @@
       // event.target: element that received focus
       // event.relatedTarget: element that lost focus
     }, true);
-
- */
+*/
 
 import '../prototype/window.customevent';
+import decorateSingleton from '../util/decorate-singleton';
 
-var previousActiveElement = document.activeElement;
+let previousActiveElement = document.activeElement;
+let raf;
 
 function observeActiveElement() {
   if (document.activeElement !== previousActiveElement) {
     // https://developer.mozilla.org/en/docs/Web/API/CustomEvent
-    var activeElementEvent = new CustomEvent('active-element', {
+    const activeElementEvent = new CustomEvent('active-element', {
       bubbles: false,
       cancelable: false,
       detail: {
@@ -36,7 +40,15 @@ function observeActiveElement() {
     previousActiveElement = document.activeElement;
   }
 
-  requestAnimationFrame(observeActiveElement);
+  raf = requestAnimationFrame(observeActiveElement);
 }
 
-observeActiveElement();
+function engage() {
+  observeActiveElement();
+}
+
+function disengage() {
+  cancelAnimationFrame(raf);
+}
+
+export default decorateSingleton({ engage, disengage });